FLEXSIM-6259 Fixed heat map exceptions
I honestly still don't know the root cause of this issue. It is hard to
replicate. On my machine it will sometimes happen when I reset the model
or open a new model. I tried resizing the grid dynamically and couldn't
it to happen, at least until I reset or opened a new model, and it would
quietly crash.

- In the case that it's a buffer overrun issue, I combed the code up and
down to find some code issue, but couldn't. Nevertheless, just in case,
I added 20 extra pixels to the end of the buffer. I also put various other
protections around it. Even though after making these changes it didn't
appear to fix it completely, I'm going to leave these in because maybe
they're good checks to have in any case.

- I also wondered if it might be the fact that I'm creating and deleting a
new texture on every repaint here. So, instead of creating/deleting
every time, I store a textureID on the Grid, which references the created
texture. I delete the texture on reset and in the Grid's destructor. After
making this change, I could no longer get FlexSim to crash. So, I'm going
to assume this is the issue. Maybe it's graphics card specific or something,
where some graphics cards (or all?) don't like you creating and deleting
textures so often. Who knows.