FLEXSIM-5727 Cut off grid rendering at a million entries
If the user accidentally creates a barrier out in the middle of
nowhere, it can cause the computer to freeze in trying to
allocate the main memory and graphics memory need to
render the entire grid. The "better" solution is to use instanced
rendering so you don't have to store so much for each grid
entry. But for now, I'm just going to limit the number of grid
entries at a million. User will hopefully figure out something is
wrong and fix it at that point.