FLEXSIM-7288 Fixed mandatory path not being applied.
Commit 65aff35 changed how temporary barriers are applied.
Mandatory paths seem to be applied as conditional barriers
even without a condition, which messes up the assumptions
in that commit.

This fix sets the isOnMandatoryPath bit to true for mandatory
path node values whether there's a conditional rule or not so
that it gets applied down correctly in AStarNavigator::calculatePath()
even with the change to how temporary barriers are applied.