Fixing Bugs
https://test.flexsim.com/index.php?/tests/view/53441&group_by=cases:section_id&group_id=1171&group_order=asc
https://test.flexsim.com/index.php?/tests/view/53500&group_by=cases:section_id&group_id=1171&group_order=asc
1. I found a bug in the collision avoidance algorithm (that's been there all along) where
a traveler can "miss" the creation of a blocking event, so it drives on through. And for good
measure I adjusted the deadlock recovery algorithm to randomly shuffle the search order
for getting out of the deadlock when the traveler encounters multiple deadlocks in the
same travel operation. This means it will be less likely to get into the loopy deadlock thing.
2. I fixed mandatory paths to be properly one-way.