Fixed issue with objects rotating regardless of TE no-rotate setting
If the user checks the box "Stop for Turns" on the A* navigator properties
then the AGV will always rotate, even if the user has said unchecked the
"Rotate while traveling" box. I changed it so that, if the user unchecks
that box, it will simulate the delay associated with changing direction,
without actually rotating the object. This came up in my AutoStore model.