Fixed crash when travel path has multiple traverrsals of same cell
The problem was that when the traveler clears his allocations
(Traveler.cpp:652), it inside a list iteration. If the cleared allocations
include the 'next' iterator in the list iteration, then it will crash
because the iterator now points to a removed entry.

So, instead I pass a reference to the iterator into the predicate, and
if the predicate removes stuff from the list, then it is responsible for
updating the iterator properly before removing stuff.

I also updated navigatePath() to properly update the index update when
multiple allocations are removed.

https://answers.flexsim.com/questions/151304/bug-report-flexsim-2321-crashing-due-to-a-collisio.html