Fixing A* issues
In figuring out the start cell from which to navigate around deadlock, it
should offset the calculation of atDist a tiny bit

Also, another previous change broke the kiva system test model, so I
needed to adjust the logic when startingBetweenCells, if the guy is
practically on a node, might as well just use that node and set
startBetweenCells to false.

And I did some UI/tooltip fixes.