FLEXSIM-8811 Showing more information for debug
The user added the flow item bin grid bot as an A* member, which was
causing it to allocate a cell, blocking other grid bots. So, kind of
user error, but we should put things in place to, first, better help
figure out what the problem is, and second, prevent the user from doing
this.

I'm not sure how we should go about the prevention part (I don't like
putting a special rule in the code just for this case). For now, I'm
dealing with it through better information display. I make it so when the
user hovers over an allocation, it displays the owning traveler and time
range of allocation. So the user, if they show allocations, will see that
the cell is allocated, and see that the owner is the robot in the flow
item bin.