FLEXSIM-8831 Don't allocate nodes for a flowitem bin TE
I don't like special rules like 'if you are in the flow item bin, don't
add.' But I do like a more generalized rule like 'if you are not in the
model's scene graph, don't add.' This test checks if there is a non-object-data
parent/ancestor of the TE, (e.g. the flow item bin folder). If so, then since
that object is not part of the model scene graph (it won't be shown in
the 3D view), I shouldn't allocate grid nodes.