FLEXSIM-8749 Changed ordering of removals in clearAllocations()
The following four configurations combine to cause the issue:
1. removing an allocation can cause another blocked traveler to immediately
  continue
2. A traveler's allocations deque is forward-time-ordered
3. clearAllocations() removes allocations in from start to end (in forward-
  time-order)
4. In Traveler::abortTravel(), the traveler deletes its blockEvent, and
  THEN calls clearAllocations().

This means that, as clearAllocations() removes each allocation, another
traveler could be continued, which can cause it to recognize future allocation
collisions with the same traveler that is currently clearing its allocations,
and subsequently recreate the blockEvent on that traveler, right after it
deleted the blockEvent. So you then have a 'dangling' blockEvent that later
references an invalid index in travelPath because travelPath has been reset,
and then all hell breaks loose.

To fix it, traverse allocations in revere-time-order in clearAllocations(),
so that when the other traveler is continued, all future allocations on this
traveler will have been removed.