Made editing spatials/loc in visuals panel work.
Edits are saved off in resizeinfo node onPreDraw and are referenced in
drawBounds. UpdateSpatials was changed to accept a bool (setMode) which
if true will set the grid min/maxpoints to the newly-edited spatials,
instead of the other way around. Default behavior for updatespatials
(bool false) now sets spatial dimensions to the bounds vertices.
Made it so color edit works for Grid.