Fixed repeatability issue
The problem was that the toEntry.turnEndTime is technically set to
the time that the to-next cell movement starts, meaning there may be a
gap where the code gets into this if statement after the object has
finished rotating to the destination direction. So the object may
over-rotate. So I just put bounds on the rotation. If the turnDist is
less than 0 (shouldn't be but just to make sure) it will set it to the
fromAngle, if the turnDist is positive but less than the max turn dist,
set it to the resolved rotation, and otherwise set it to toAngle.