Fix allocation release time adjustment in onBlock
FLEXSIM-11163 When an AGV is blocked, it has to decelerate to a stop.
Because it begins slowing down, it should extend its allocations to
account for its slower speeds. The issue is that some allocations are
only being held because of the navigator's deallocTimeOffset. Those
allocations should not be extended.